\name{BayesOrd-package}
\alias{BayesOrd-package}
\alias{BayesOrd}
\docType{package}
\title{
MCMC routines for model fitting and Bayesian model choice for ordinal regression models
}
\description{
Package implements routines described in McKinley et al. (2014), in review, for fitting ordinal regression models in a Bayesian framework using Markov chain MOnte Carlo methods. The package can be used to fit proportional odds (PO) and non-proportional odds (NPO) models, and can also be used to perform Bayesian model selection to choose between PO or NPO structures for each variable. In addition the package will perform variable selection based on posterior probabilities of association.
}
\details{
\tabular{ll}{
Package: \tab BayesOrd\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-01-28\cr
License: \tab GPL-3\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
TJ McKinley

Maintainer: TJ McKinley <tjm44@cam.ac.uk>
}
\references{McKinley et al., (2014), in review.}
\seealso{\code{\link{bayesord}}}
