\name{fitted.bayesord}
\alias{fitted.bayesord}
\alias{print.fitted.bayesord}
\alias{summary.fitted.bayesord}
\alias{print.summary.fitted.bayesord}
\alias{plot.fitted.bayesord}
\title{Predictive posterior distributions for observed data, obtained from model outputs}
\description{
Produces predictive posterior distributions for observed data, based on output
generated by a call to \code{\link{bayesord}}.
}
\usage{
\method{fitted}{bayesord}(object, multi = F, mc.cores = NA, cluster = TRUE, ...)
\method{summary}{fitted.bayesord}(object, ...)
\method{print}{fitted.bayesord}(x, ...)
\method{plot}{fitted.bayesord}(x, variable = "all", ...)
}
\arguments{
  \item{object}{an object of class "bayesord" (for \code{fitted}), or class "fitted.bayesord" (for \code{summary}).}
  \item{x}{an object of class "fitted.bayesord".}
  \item{multi}{a logical specifying whether to run the code in parallel.}
  \item{mc.cores}{an integer specifying how many cores to use for parallel processing.}
  \item{cluster}{a logical specifying whether 'cluster' terms were specified in the model.}
  \item{variable}{a character determining which variable to plot from model/data. Can take the generic value \code{"all"} to denote all variables.}
  \item{\dots}{ignored here.}
}
\details{
'print.fitted.bayesord' prints summary of classification probabilities based on predictive posterior distributions and the observed data.

'plot.fitted.bayesord' plots marginal predictive posterior distributions against the observed data.

In order for the \code{multi} argument to be used, the \code{multicore} package must be installed. If \code{mc.cores} is not specified, then it defaults to the number of available cores on the machine.
}
\value{
Produces a list with elements:
\item{fits}{a list, where each element of the list is an (ngroup x niter)-matrix corresponding to each row of 'model.dat'. Each matrix contains predicted posterior samples for that data configuration.}
\item{model.dat}{a data frame where each row corresponds to a unique configuration of data points (used for improving efficiency).}
\item{var.info}{a list, containing information relating to the original data. This is used to produce the marginal plots.}
}
\references{McKinley et al., (2014), to appear in Bayesian Analysis.}
\author{TJ McKinley}
\seealso{\code{\link{bayesord}}}
\examples{
# load in US data on religious fundamentalism against region of residence
# (data can be found in Agresti, "Analysis of Ordinal Categorical Data", (2010))
data(religion)

# fit non-proportional odds model
religion.NPO <- bayesord(Beliefs ~ Region, data = religion, niter = 2e+05, noutputsum = 2e+04, 
    nchains = 2, model.type = "NPO", start = 50000, thin = 150)
    
## check convergence plots and summarise output (NOT RUN)
#plot(religion.NPO)
#print(religion.NPO)
#summary(religion.NPO)

# examine fitted model
fitted.NPO <- fitted(religion.NPO)
# summarise and plot predictive posteriors
print(fitted.NPO)
plot(fitted.NPO)
}
