\name{mentalhealth}
\alias{mentalhealth}
\docType{data}
\title{Mental health against life events and socio-economic status}
\description{These data come from a study of mental health in a sample of 40 adults living in Alachua County, Florida, US. More details can be found in Agresti (2010), P62.
}
\usage{data(mentalhealth)}
\format{Data frame consisting of three columns:  "Mental.Health"---a factor with four categories: 'Well', 'Mild', 'Moderate' and 'Impaired'; "SES"---a binary vector; and "Life.Events"---a discrete vector containing a composite measure of the number and severity of
different life events.}
\source{Agresti, "Analysis of Ordinal Categorical Data (2nd Edition)", Wiley, (2010), P62}

