\name{religion}
\alias{religion}
\docType{data}
\title{Cross-classification table of religious beliefs against region of residence in the United States}
\description{
These data are from the 2006 General Social Survey in the United States and are provided in Agresti (2010), P72. The data consist of a cross-classification table of religious belief against the region in which the respondant lived. 
}
\usage{data(religion)}
\format{Data frame consisting of two columns: "Belief"---a factor with three categories: 'fundamentalist', 'moderate' and 'liberal'; and "region"---a factor with four categories: 'Northeast', 'Midwest', 'South' and 'West/Mountain'.}
\source{Agresti, "Analysis of Ordinal Categorical Data (2nd Edition)", Wiley, (2010), P72}

